pyproject.toml
[project]
authors = [
  {name = "Mrowr Purr", email = "mrowr.purr@gmail.com"},
]
dependencies = ["pyqt6 (>=6.9.0,<7.0.0)"]
description = "Mod Organizer 2 - Plugin Examples"
license = {text = "GPL-3.0"}
name = "python-examples"
readme = "README.md"
requires-python = ">=3.12,<3.14"
version = "0.1.0"

[tool.poetry.group.dev.dependencies]
mobase-stubs = "^2.5.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[tool.pyright]
exclude = ["**/__pycache__", "**/.venv", "**/venv", "**/.mypy_cache"]
reportCallInDefaultInitializer = true
reportImplicitOverride = true
reportImplicitStringConcatenation = true
reportImportCycles = true
reportMissingModuleSource = false
reportMissingSuperCall = false
reportPropertyTypeMismatch = true
reportShadowedImports = true
reportUninitializedInstanceVariable = true
reportUnnecessaryTypeIgnoreComment = true
reportUnusedCallResult = false
strictDictionaryInference = true
strictListInference = true
strictSetInference = true
typeCheckingMode = "strict"

[tool.black]
line-length = 180
